Run tests in docker container.
