+++ tests/sanity/run.py
@@ -1,64 +1,31 @@
 #!/usr/bin/env python
 
+# Copyright: (c) 2020, Felix Fontein <felix@fontein.de>
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 
+import os
 import subprocess
 import sys
 
 
-SEPARATOR = '=========================================================================='
-
-
-def run(failed_tests, errors, test, command):
-    try:
-        p = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        stdout, stderr = p.communicate()
-
-        new_errors = []
-        if stdout:
-            stdout = stdout.decode('utf-8').splitlines()
-            for line in stdout:
-                new_errors.append((test, '{0}'.format(line)))
-        if stderr:
-            stderr = stderr.decode('utf-8').splitlines()
-            for line in stderr:
-                new_errors.append((test, '[stderr] {0}'.format(line)))
-        if int(p.returncode) != 0:
-            new_errors.append((test, '[command returned {0}]'.format(p.returncode)))
-    except Exception as e:
-        new_errors.append((test, '[internal error] {0}'.format(e)))
-
-    if new_errors:
-        print(SEPARATOR)
-        print('Test "{0}" failed with the following {1} errors:'.format(test, len(new_errors)))
-        failed_tests.append(test)
-        for dummy, line in new_errors:
-            print(line)
-        errors.extend(new_errors)
-    else:
-        print(SEPARATOR)
-        print('Test "{0}" succeeded.'.format(test))
-
-
 def main():
-    errors = []
-    failed_tests = []
+    root = os.environ['HOME']
 
-    run(failed_tests, errors, 'changelog', ['ansible-changelog', 'lint'])
-    run(failed_tests, errors, 'bundled', ['tests/sanity/code-smell/update-bundled.py', 'plugins/module_utils/compat/ipaddress.py'])
-
-    if not errors:
-        print(SEPARATOR)
-        sys.exit(0)
-
-    print(SEPARATOR)
-    print('Total of {0} errors in the following tests:'.format(len(errors)))
-    for test in failed_tests:
-        print(test)
-    print(SEPARATOR)
-    sys.exit(-1)
+    command = ['docker', 'run', '--rm', '-t']
+    command.extend(['-v', '{0}:{1}'.format(root, root)])
+    command.extend(['-w', '/{0}'.format(os.getcwd())])
+    command.extend(['-u', '{0}:{1}'.format(os.getuid(), os.getgid())])
+    command.extend(['quay.io/ansible/default-test-container:{0}'.format('1.12')])
+    # command.extend(['/bin/sh', '-c', 'ls -lah ; pwd'])
+    command.extend(['python3.7', 'tests/sanity/runner.py'])
+    command.extend(sys.argv[1:])
+    print('[RUN] {0}'.format(' '.join(command)))
+    rc = subprocess.call(command)
+    sys.exit(rc)
 
 
 if __name__ == '__main__':
